util/cbfstool: Use `%zu` instead of `%ld` for size_t arguments
cbfstool fails to built under 32-bit platforms since commit

    aa2f739a cbfs: fix issues with word size and endianness.

due to the use of '%ld' format specifier on size_t, which on these
platforms is only 32-bit.

No error is seen though, when cbfstool is built, when building a coreboot
image, where it is put in `build/cbfstool`.

Use the length modifier `z` for size_t arguments, and cast to size_t where
appropriate.

Reviewed-on: http://review.coreboot.org/5388
Tested-by: build bot (Jenkins)

+++ util/cbfstool/cbfs-mkstage.c
@@ -132,8 +132,8 @@ int parse_elf_to_stage(const struct buffer *input, struct buffer *output,
 			continue;
 		if (input->size < (phdr[i].p_offset + phdr[i].p_filesz)){
 			ERROR("Underflow copying out the segment."
-			      "File has %ld bytes left, segment end is %ld\n",
-			      input->size, phdr[i].p_offset + phdr[i].p_filesz);
+			      "File has %zu bytes left, segment end is %zu\n",
+			      input->size, (size_t)(phdr[i].p_offset + phdr[i].p_filesz));
 			return -1;
 		}
 		memcpy(buffer + (l_start - data_start),
